separated SortedFileList & FileSortingMode, FileData & FileType, FileRecyclerViewHolder & FileRecyclerAdapter
